+++ lib/opro/controllers/concerns/error_messages.rb
@@ -6,6 +6,7 @@ module Opro::Controllers::Concerns::ErrorMessages
     msg << ' - No OAuth Token Provided!'    if oauth_access_token.blank?
     msg << ' - Allow OAuth set to false!'   if allow_oauth? == false
     msg << ' - OAuth user not found!'       if oauth_user.blank?
+    msg << ' - OAuth client has been rate limited' if oauth_client_over_rate_limit?
     msg = generate_oauth_permissions_error_message!(msg)
     msg
   end